API-12088 remove common objects from customer and adjust structures
